header menu refactor and add community links loading
